Refactor read_evidence module into subpackage
Also address a few minor issues raised in code review.
